requirements.txt
pygame==2.6.1
numpy==1.24.3
matplotlib==3.7.1
# Optional dependencies
openai>=1.3.0; python_version >= "3.9"
python-dotenv>=1.0.0; python_version >= "3.9"
requests>=2.31.0; python_version >= "3.9"
tiktoken==0.5.1